refactor: clarify legacy backup envelope naming
